Fixed Issue 218 (http://code.google.com/p/beast-mcmc/issues/detail?id=218).

+++ src/dr/app/beast/BeastMain.java
@@ -389,7 +389,7 @@ public class BeastMain {
 
             BeastDialog dialog = new BeastDialog(new JFrame(), titleString, icon);
 
-            if (!dialog.showDialog(nameString, (int)seed)) {
+            if (!dialog.showDialog(nameString, seed)) {
                 return;
             }
